requirements.txt
groq>=0.4.0
elevenlabs>=0.2.0
python-dotenv>=1.0.0
obs-websocket-py>=0.6.0